requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
arabic-reshaper==3.0.0
asgiref==3.9.1
asn1crypto==1.5.1
attrs==25.3.0
black==22.12.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
crispy-bootstrap5==0.7
cryptography==45.0.6
cssselect2==0.8.0
Deprecated==1.2.18
dj-database-url==3.0.1
Django==5.2.6
django-ckeditor==6.7.3
django-ckeditor-5==0.2.18
django-crispy-forms==1.14.0
django-extensions==3.2.3
django-filter==23.5
django-jazzmin==3.0.1
django-jet-reboot==1.3.5
django-js-asset==3.1.2
django-model-utils==4.3.1
django-modeltranslation==0.18.11
django-pwa==2.0.1
django-widget-tweaks==1.5.0
factory_boy==3.3.1
Faker==37.5.3
frozenlist==1.7.0
gopay==2.0.1
gunicorn==23.0.0
html5lib==1.1
idna==3.10
lxml==6.0.1
multidict==6.6.4
mypy_extensions==1.1.0
oscrypto==1.3.0
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
propcache==0.3.2
psycopg2==2.9.10
pycparser==2.22
pydantic==1.10.22
pyHanko==0.30.0
pyhanko-certvalidator==0.28.0
PyJWT==2.10.1
pypdf==6.0.0
python-bidi==0.6.6
python-decouple==3.8
python-docx==1.2.0
pytz==2022.7
PyYAML==6.0.2
reportlab==4.0.4
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
stripe==5.5.0
svglib==1.5.1
tinycss2==1.4.0
twilio==9.7.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
webencodings==0.5.1
whitenoise==6.2.0
wrapt==1.17.3
xhtml2pdf==0.2.15
yarl==1.20.1
